+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 [Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-grafana/compare/0.1.1...HEAD
 
+### Changed
+
+-   Requires chaostoolkit-lib 1.26.0 at least
+
 ## [0.1.1][]
 
 [0.1.1]: https://github.com/chaostoolkit-incubator/chaostoolkit-grafana/compare/0.1.0...0.1.1